Fixed accural type mismatch.

+++ internal/accrual/updater.go
@@ -12,9 +12,9 @@ import (
 )
 
 type orderInfo struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int64  `json:"accrual"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual"`
 }
 
 type Updater struct {
